+++ src/senaite/fhir/converter/analysisrequest.py
@@ -81,6 +81,9 @@ class ResourceToAnalysisRequest(object):
         services = self.get_services()
         priority = self.get_priority()
 
+        external_id = self.resource.get_external_id()
+        client_sample_id = external_id.value if external_id else None
+
         data = {
             "portal_type": "AnalysisRequest",
             "parent_path": api.get_path(client),
@@ -89,12 +92,13 @@ class ResourceToAnalysisRequest(object):
             "SampleType": sample_type,
             "SamplePoint": sample_point,
             "DateSampled": date_sampled,
-            "Profiles": [profile] if profile else [],
+            "Profiles": [api.get_uid(profile)] if profile else [],
             "Priority": priority,
             # "Sampler": collector,
             # "Remarks": remarks,
             "Specification": specs,
             "Analyses": services,
+            "ClientSampleID": client_sample_id,
         }
 
         # update with patient information